test: address runtime troubleshooting review
