+++ e2e/runtime-register/fixtures/exec-argv.test.ts
@@ -1,12 +1,6 @@
-import { rmSync } from 'node:fs';
 import { expect, test } from '@rstest/core';
 
-test('passes pool.execArgv node flags to workers', ({ onTestFinished }) => {
-  const registerFlagPath = process.env.RUNTIME_REGISTER_FLAG_PATH;
-  if (registerFlagPath) {
-    onTestFinished(() => rmSync(registerFlagPath, { force: true }));
-  }
-
+test('passes pool.execArgv node flags to workers', () => {
   expect(process.execArgv).toContain('--conditions=rstest-e2e');
   expect(process.execArgv).toContain('--require');
   expect(process.execArgv).toContain('./cjs-register.cjs');
